requirements.txt
streamlit
python-dotenv
openai
pydantic
pytest